requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.7.0
pytz==2024.1
sqlparse==0.4.4 